Moves Retrievers into a directory
