+++ src/app/retrievers/TideRetriever.cpp
@@ -7,7 +7,7 @@
 #include <utility>
 
 #include "TideRetriever.h"
-#include "Time.h"
+#include "src/app/Time.h"
 
 std::string TideRetriever::Retrieve() {
     std::cout << "TideRetriever::Retrieve" << std::endl;
